Setup construct URL (for now without logic, i.e. just with hard coding)
We're going to create two helper functions which are going to be useful.
Essentially we create a function which is going to return the URL where we need to navigate.
We'll do multiple things in those helper functions. First, we'll grab all of the existing query params again, something we'll probably repeat multiple times
during the pagination setup.
And after that we will add the page where we want to go, since this is something we want to provide in these query params.